requirements.txt
connexion == 2.14.2
python_dateutil == 2.8.2
setuptools == 75.8.0
swagger-ui-bundle == 0.0.9
boto3==1.37.11
flask==2.2.5
pyjwt
requests
